VABGAtuning
The methods for testing the GA in various configurations have been
moved out of Test_VABProcess.py and into VABGAtuning. This is so
that the other tests can be executed quickly with nose.
Also, the GA in VABProcess has been augmented with a reporting
file called "out" added to whatever directory the code is executing in.