Another dead end
This commit includes a procedure in which random subsampling is used to
estimate a distribution for the coefficient of determination, and these
distributions are compared using a Kolmogorov-Smirnoff 2-sample test.
However, the very same system seems to yield a data set that, when
subsampled, gives a distinct distribtion of R2 values. A different
approach is needed.