Successful use of F-test method
This commit includes some modifications to the base VAB classes, as well
as a new method in VABProcess that compares models (or, more accurately,
the data behind them) using an F-test. The method is incomplete
(currently it only returns a list of p-values), but it works for modest
noise levels (zero noise causes problems, but I suspect that's because
of violations of the normality assumption).