New directory structure
Created a directory "eugene" to hold the package containing all elements
of the automated discovery routine. The directory "assess" is also
intended to hold a Python package (though not for public dissemination).
Finally, I have added a "scripts" directory to hold all of our
experiment, testing, and figure production scripts. This is not set up
as a package.