Revised core algorithm
The CompareModels method for EUGENE has been rewritten to dispense with
the F-test in favor of nested cross-validation routines. However, the
resutls is, somewhat mysteriously, overly sensitive to noise in the
data. This is clearly a fundamental problem that needs to be solved.
It's possible that I am not comparing apples to apples with the test as
currently constructed.

Other important changes:
- I corrected the cross-validation routines used in BuildSymModel and
  elsewhere. They were originally taking the mean of the mean square
  error for different curves rather than summing all errors and then
  taking a mean.

- I added two files for testing: TestCompare.py and CLMPSdemo.py