Bug hunting
Cleaned up a few errors (empty ndarrays do not behave as one might
expect) and standardized the use of the "residuals" function for
computing fit errors. However, it is still the case the when running
GrowthDemo some comparisons still indicated a much HIGHER error for
separate fits than for joint fits.