Bug partially isolated
I've tracked the remaining bug to the calls made to the 'residuals'
methods within the testing phase of CompareModels. Those are the only
calls which pass ydata as a column vector rather than a flat array. That
doesn't explain why only some of the calls are bad (those for the
base_error seperate models), but it's a start. This commit proceeds a
revision intended to normalize these calls.