Module for converting distance matrices to maps
Added a module in tools that takes a distance matrix, uses it as a
similarity matrix for MDS, and that treats the MDS coordinates as points
in color space to color a map in parameter space.

Note: also made a minor tweak to the documentation in LVDSim without
realizing what branch I was on.