Bug fixes to ensure all tests are satisfied
Code for mutliple modules adjusted to ensure that all tests in the tests
folder complete successfully when run under Python 3.6. Most fixes had
to do with the fact that Python 2 and 3 cast integers differently.
However, one fix had to do with a new warning added to the method for
choosing untrans and trans curves.

Note that test_robustness.py was moved out of this repository.

 On branch develop
 Changes to be committed:
	modified:   LICENSE.txt
	modified:   eugene/src/data_prep/initial_conditions.py
	modified:   eugene/src/data_prep/metaparameters.py
	modified:   eugene/src/dd_scan.py
	modified:   eugene/tests/test_initial_conditions.py
	deleted:    eugene/tests/test_robustness.py

 Untracked files:
	pyproject.toml
	setup.cfg